+++ R/ichimoku-package.R
@@ -81,13 +81,13 @@
 #'     PositionIdentity rel scale_color_manual scale_fill_manual
 #'     scale_x_continuous scale_y_continuous Stat StatIdentity theme theme_grey
 #'     %+replace%
-#' @importFrom jsonlite parse_json
 #' @importFrom nanonext ncurl sha256
 #' @importFrom shiny checkboxInput column downloadButton downloadHandler HTML
 #'     fillPage fluidPage fluidRow hoverOpts invalidateLater isolate
 #'     numericInput observeEvent plotOutput reactive reactiveVal renderPlot
 #'     renderUI req selectInput shinyApp sliderInput stopApp tags textInput
 #'     uiOutput wellPanel
+#' @importFrom RcppSimdJson fparse
 #' @importFrom stats na.omit sd
 #' @importFrom utils packageVersion str
 #' @importFrom xts endpoints